+++ HaloEzAPI/Abstraction/Interfaces/IApiCacheManager.cs
@@ -1,0 +1,11 @@
+﻿namespace HaloEzAPI.Abstraction.Interfaces
+{
+    public interface IApiCacheManager
+    {
+        void Add<T>(T obj, string key, int expiryHours) where T : class;
+        T Get<T>(string key) where T : class;
+        bool Contains(string key);
+        void Remove(string key);
+        void RemoveAll();
+    }
+}