+++ HaloEzAPI/Model/Response/Stats/StatTimelapse.cs
@@ -7,7 +7,6 @@ namespace HaloEzAPI.Model.Response.Stats
     public class StatTimelapse
     {
         public Guid Id { get; set; }
-        [JsonIgnore]
         [JsonConverter(typeof(TimeSpanConverter))]
         public TimeSpan Timelapse { get; set; }
     }